refactor: Introduce FrameNames class and update FRAME_STORE references
