+++ app/frames/copy_flowline_data.py
@@ -4,7 +4,8 @@ import tkinter as tk
 from tkinter import messagebox, ttk
 
 from app.core.multi_case_modeller import copy_flowline_data
-from app.project import FRAME_STORE, browse_folder_or_file
+from app.frames import FRAME_STORE, FrameNames
+from app.project import browse_folder_or_file
 
 logger_uc = logging.getLogger("app.core.simulation_modeller")
 
@@ -31,7 +32,7 @@ def submit_copy_flowline_data(
 
 def init_update_conditions_frame(app: tk.Tk) -> tk.Frame:
     update_conditions_frame = tk.Frame(app)
-    FRAME_STORE["update_conditions"] = update_conditions_frame
+    FRAME_STORE[FrameNames.COPY_FLOWLINE_DATA] = update_conditions_frame
     update_label = tk.Label(
         update_conditions_frame, text="Copy Flowline Data Workflow", font=("Arial", 14)
     )
